Use interfaces from web for api as well; install express-session types
