+++ api/package.json
@@ -50,6 +50,7 @@
 		"@babel/preset-env": "^7.1.0",
 		"@babel/preset-typescript": "^7.1.0",
 		"@types/express": "^4.16.0",
+		"@types/express-session": "^1.15.11",
 		"@types/googlemaps": "^3.30.16",
 		"@types/mysql": "^2.15.5",
 		"@types/node": "^10.11.7",
